refactory the signal, process, pprof to independent folders.
