+++ comet/config.go
@@ -20,6 +20,7 @@ import (
 	"errors"
 	"flag"
 	"github.com/Terry-Mao/goconf"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"runtime"
 	"time"
 )
