--- comet/pprof.go
@@ -1,40 +1,0 @@
-// Copyright © 2014 Terry Mao, LiuDing All rights reserved.
-// This file is part of gopush-cluster.
-
-// gopush-cluster is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// gopush-cluster is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with gopush-cluster.  If not, see <http://www.gnu.org/licenses/>.
-
-package main
-
-import (
-	"net/http"
-	"net/http/pprof"
-)
-
-// StartPprof start http pprof.
-func StartPprof() {
-	pprofServeMux := http.NewServeMux()
-	pprofServeMux.HandleFunc("/debug/pprof/", pprof.Index)
-	pprofServeMux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
-	pprofServeMux.HandleFunc("/debug/pprof/profile", pprof.Profile)
-	pprofServeMux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
-	for _, addr := range Conf.PprofBind {
-		go func() {
-			Log.Info("start pprof listen addr:\"%s\"", addr)
-			if err := http.ListenAndServe(addr, pprofServeMux); err != nil {
-				Log.Error("http.ListenAdServe(\"%s\") error(%v)", addr, err)
-				panic(err)
-			}
-		}()
-	}
-}
