+++ comet/pubsub.go
@@ -18,6 +18,7 @@ package main
 
 import (
 	"errors"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"time"
 )
