+++ comet/pubsub_tcp.go
@@ -19,6 +19,7 @@ package main
 import (
 	"bufio"
 	"errors"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"io"
 	"net"
 	"strconv"
