+++ comet/pubsub_websocket.go
@@ -18,6 +18,7 @@ package main
 
 import (
 	"code.google.com/p/go.net/websocket"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"net"
 	"net/http"
 	"strconv"
