+++ comet/zk.go
@@ -19,6 +19,7 @@ package main
 import (
 	"errors"
 	"fmt"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"launchpad.net/gozk/zookeeper"
 	"strings"
 )
