+++ message/config.go
@@ -20,6 +20,7 @@ import (
 	"flag"
 	"fmt"
 	"github.com/Terry-Mao/goconf"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"runtime"
 	"time"
 )
