+++ message/main.go
@@ -18,18 +18,18 @@ package main
 
 import (
 	"flag"
-	"github.com/Terry-Mao/gopush-cluster/log"
+	. "github.com/Terry-Mao/gopush-cluster/log"
+	. "github.com/Terry-Mao/gopush-cluster/pprof"
+	. "github.com/Terry-Mao/gopush-cluster/process"
+	. "github.com/Terry-Mao/gopush-cluster/signal"
 	"os"
 	"runtime"
 	"time"
 )
 
-var (
-	Log = log.DefaultLogger
-)
-
 func main() {
 	var err error
+	Log = DefaultLogger
 	// Parse cmd-line arguments
 	flag.Parse()
 	signalCH := InitSignal()
@@ -45,7 +45,7 @@ func main() {
 	runtime.GOMAXPROCS(Conf.MaxProc)
 
 	// Load log
-	Log, err = log.New(Conf.LogFile, Conf.LogLevel)
+	Log, err = NewLog(Conf.LogFile, Conf.LogLevel)
 	if err != nil {
 		panic(err)
 		os.Exit(-1)
@@ -54,13 +54,13 @@ func main() {
 	// init process
 	// sleep one second, let the listen start
 	time.Sleep(time.Second)
-	if err = InitProcess(); err != nil {
+	if err = InitProcess(Conf.User, Conf.Dir, Conf.PidFile); err != nil {
 		Log.Error("InitProcess() error(%v)", err)
 		os.Exit(-1)
 	}
 
 	// start pprof http
-	StartPprof()
+	StartPprof(Conf.PprofBind)
 
 	// Initialize redis
 	InitRedis()
