+++ message/redis.go
@@ -20,6 +20,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/Terry-Mao/gopush-cluster/hash"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"github.com/garyburd/redigo/redis"
 )
