+++ signal/signal.go
@@ -14,27 +14,36 @@
 // You should have received a copy of the GNU General Public License
 // along with gopush-cluster.  If not, see <http://www.gnu.org/licenses/>.
 
-package main
+package signal
 
 import (
-	"net/http"
-	"net/http/pprof"
+	. "github.com/Terry-Mao/gopush-cluster/log"
+	"os"
+	"os/signal"
+	"syscall"
 )
 
-// StartPprof start http pprof.
-func StartPprof() {
-	pprofServeMux := http.NewServeMux()
-	pprofServeMux.HandleFunc("/debug/pprof/", pprof.Index)
-	pprofServeMux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
-	pprofServeMux.HandleFunc("/debug/pprof/profile", pprof.Profile)
-	pprofServeMux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
-	for _, addr := range Conf.PprofBind {
-		go func() {
-			Log.Info("start pprof listen addr:\"%s\"", addr)
-			if err := http.ListenAndServe(addr, pprofServeMux); err != nil {
-				Log.Error("http.ListenAndServe(\"%s\") error(%v)", addr, err)
-				panic(err)
-			}
-		}()
+// InitSignal register signals handler.
+func InitSignal() chan os.Signal {
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, syscall.SIGHUP, syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT, syscall.SIGSTOP)
+	return c
+}
+
+// HandleSignal fetch signal from chan then do exit or reload.
+func HandleSignal(c chan os.Signal) {
+	// Block until a signal is received.
+	for {
+		s := <-c
+		Log.Info("get a signal %s", s.String())
+		switch s {
+		case syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGSTOP, syscall.SIGINT:
+			return
+		case syscall.SIGHUP:
+			// TODO reload
+			//return
+		default:
+			return
+		}
 	}
 }
