+++ web/handle.go
@@ -19,6 +19,7 @@ package main
 import (
 	"encoding/json"
 	"fmt"
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	"io"
 	"net/http"
 	"strconv"
