+++ web/main.go
@@ -18,7 +18,10 @@ package main
 
 import (
 	"flag"
-	"github.com/Terry-Mao/gopush-cluster/log"
+	. "github.com/Terry-Mao/gopush-cluster/log"
+	. "github.com/Terry-Mao/gopush-cluster/pprof"
+	. "github.com/Terry-Mao/gopush-cluster/process"
+	. "github.com/Terry-Mao/gopush-cluster/signal"
 	"net"
 	"net/http"
 	"os"
@@ -28,10 +31,6 @@ import (
 
 const (
 	httpReadTimeout = 30 //seconds
-)
-
-var (
-	Log *log.Logger
 )
 
 func main() {
@@ -51,7 +50,7 @@ func main() {
 	runtime.GOMAXPROCS(Conf.MaxProc)
 
 	// Load log
-	Log, err = log.New(Conf.LogPath, Conf.LogLevel)
+	Log, err = NewLog(Conf.LogPath, Conf.LogLevel)
 	if err != nil {
 		panic(err)
 		os.Exit(-1)
@@ -77,7 +76,7 @@ func main() {
 	}
 
 	// start pprof http
-	StartPprof()
+	StartPprof(Conf.PprofBind)
 
 	// Internal admin handle
 	go func() {
@@ -119,7 +118,7 @@ func main() {
 	// init process
 	// sleep one second, let the listen start
 	time.Sleep(time.Second)
-	if err = InitProcess(); err != nil {
+	if err = InitProcess(Conf.User, Conf.Dir, Conf.PidFile); err != nil {
 		Log.Error("InitProcess() error(%v)", err)
 		os.Exit(-1)
 	}
