+++ web/msg_svr.go
@@ -17,6 +17,7 @@
 package main
 
 import (
+	. "github.com/Terry-Mao/gopush-cluster/log"
 	myrpc "github.com/Terry-Mao/gopush-cluster/rpc"
 	"net/rpc"
 	"time"
