--- web/signal.go
@@ -1,48 +1,0 @@
-// Copyright © 2014 Terry Mao, LiuDing All rights reserved.
-// This file is part of gopush-cluster.
-
-// gopush-cluster is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// gopush-cluster is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with gopush-cluster.  If not, see <http://www.gnu.org/licenses/>.
-
-package main
-
-import (
-	"os"
-	"os/signal"
-	"syscall"
-)
-
-// InitSignal register signals handler.
-func InitSignal() chan os.Signal {
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, syscall.SIGHUP, syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT, syscall.SIGSTOP)
-	return c
-}
-
-// HandleSignal fetch signal from chan then do exit or reload.
-func HandleSignal(c chan os.Signal) {
-	// Block until a signal is received.
-	for {
-		s := <-c
-		Log.Info("get a signal %v", s)
-		switch s {
-		case syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGSTOP, syscall.SIGINT:
-			return
-		case syscall.SIGHUP:
-			// TODO reload
-			// return
-		default:
-			return
-		}
-	}
-}
